+++ src-pwa/custom-service-worker.js
@@ -3,3 +3,16 @@
  * is picked up by the build system ONLY if
  * quasar.conf > pwa > workboxPluginMode is set to "InjectManifest"
  */
+// eslint-disable-next-line no-undef
+
+if (self.workbox) {
+  self.workbox.precaching.precache(self.__precacheManifest)
+
+  self.addEventListener('fetch', function(event) {
+    event.respondWith(
+      caches.match(event.request).then(function(response) {
+        return response || fetch(event.request)
+      })
+    )
+  })
+}