+++ src/util/common.js
@@ -75,8 +75,8 @@ export function getUserType() {
 }
 
 export async function getUserFbInfo() {
-  let popup = window.open(
-    'https://www.facebook.com/v3.2/dialog/oauth?client_id=253998778647702&redirect_uri=' +
+  window.open(
+    'https://www.facebook.com/v3.2/dialog/oauth?client_id=253998778647702&response_type=token&redirect_uri=' +
       window.location.origin +
       '/fb-login-receiver.html',
     'Facebook Login',
@@ -85,13 +85,17 @@ export async function getUserFbInfo() {
   return new Promise(resolve => {
     window.addEventListener(
       'message',
-      () => {
-        popup.close()
-        window.FB.getLoginStatus(() => {
-          window.FB.api('/me', {fields: 'name,email'}, person => {
-            resolve(person)
+      ({data}) => {
+        _ax
+          .get('https://graph.facebook.com/me', {
+            params: {
+              fields: 'id,name,email',
+              access_token: data,
+            },
           })
-        })
+          .then(({data}) => {
+            resolve(data)
+          })
       },
       {once: true}
     )